Create Coin Change II.cpp
In the "Coin Change II" problem, we are given coins of various denominations and a specific amount. Our task is to determine the number of different ways we can make up that amount using the coins. Unlike other coin change problems where we need to minimize or maximize the number of coins, here we need to find the total number of possible combinations.

For instance, with coins [1,2,5] and amount 5, we have the following combinations:

1. 5 = 5
2. 5 = 2 + 2 + 1
3. 5 = 2 + 1 + 1 + 1
4. 5 = 1 + 1 + 1 + 1 + 1
Hence, there are 4 ways to make up the amount 5 using the coins [1,2,5].